Fix issue with receipt preview
