+++ app/views/spree/conekta/payments/_conekta_oxxo.html.erb
@@ -1,6 +1,7 @@
 <div class='opps'>
   <div class='opps-header'>
-    <div class="opps-reminder">Ficha digital. No es necesario imprimir.</div>
+    <div class='opps-reminder'>Ficha digital. No es necesario imprimir.</div>
+
     <div class='opps-info'>
       <div class='opps-brand'>
         <%= image_tag 'oxxopay_brand.png', alt: 'OXXOPay', class: 'summary-oxxo-image' %>
@@ -29,6 +30,7 @@
       <li>Realiza el pago correspondiente con dinero en efectivo.</li>
       <li>Al confirmar tu pago, el cajero te entregará un comprobante impreso. <strong>En el podrás verificar que se haya realizado correctamente.</strong> Conserva este comprobante de pago.</li>
     </ol>
+
     <div class='opps-footnote'>Al completar estos pasos recibirás un correo confirmando tu pago.</div>
   </div>
 </div>
