+++ app/views/spree/conekta_oxxo/receipt.html.erb
@@ -1,14 +1,19 @@
 <html>
   <head>
-    <%= stylesheet_link_tag 'spree/frontend/solidus_oxxo_pay', media: 'all', 'data-turbolinks-track': 'false' %>
-    <link href="https://fonts.googleapis.com/css?family=Open+Sans:400,600,700" rel="stylesheet">
+    <%= stylesheet_link_tag 'spree/frontend/all',
+                            media: 'all',
+                            'data-turbolinks-track': 'false' %>
+    <link href='https://fonts.googleapis.com/css?family=Open+Sans:400,600,700'
+          rel='stylesheet'>
   </head>
+
   <body>
     <%= conekta_oxxo_receipt(@payment) %>
   </body>
-  <script type="text/javascript" charset="utf-8">
-      document.addEventListener('DOMContentLoaded', function () {
-        window.print();
-      });
+
+  <script type='text/javascript' charset='utf-8'>
+    document.addEventListener('DOMContentLoaded', function () {
+      window.print();
+    });
   </script>
 </html>
